+++ contracts/Token.sol
@@ -335,7 +335,7 @@ contract Token is
      * @notice Contract version
      */
     function version() public pure virtual returns (string memory) {
-        return "2.2.0";
+        return "2.3.0";
     }
 
     // ========================================